CommandHandlerCompilerPass renamed to CommandHandlerPass and moved into DependencyInjection/Compiler
