+++ TacticianBundle.php
@@ -1,7 +1,7 @@
 <?php namespace League\Tactician\Bundle;
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use League\Tactician\Bundle\DependencyInjection\CommandHandlerCompilerPass;
+use League\Tactician\Bundle\DependencyInjection\Compiler\CommandHandlerPass;
 use League\Tactician\Bundle\DependencyInjection\TacticianExtension;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -10,7 +10,7 @@ class TacticianBundle extends Bundle
     public function build(ContainerBuilder $container)
     {
         parent::build($container);
-        $container->addCompilerPass(new CommandHandlerCompilerPass());
+        $container->addCompilerPass(new CommandHandlerPass());
     }
 
     public function getContainerExtension()
